# RentAll/RentAll.Web/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "Autogenerated API info",
    "version": "1.0.0"
  },
  "paths": {
    "/Center": {
      "get": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_ListCenters",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetCenterDto"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_CreateCenter",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "centerDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateCenterDto"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Center"
            }
          }
        }
      }
    },
    "/Center/{id}": {
      "get": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_GetCenterById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetCenterDto"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_UpdateCenter",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "centerDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateCenterDto"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Center"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_DeleteCenter",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/Center/{id}/units": {
      "get": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_ListUnitsInCenter",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetUnitDto"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_CreateUnitInCenter",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "unitDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUnitDto"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Unit"
            }
          }
        }
      }
    },
    "/Center/{id}/units/{unitId}": {
      "get": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_GetUnitById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "unitId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetUnitDto"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_UpdateUnitInCenter",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "unitId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "unitDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUnitDto"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Unit"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_DeleteUnitById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "unitId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Unit"
            }
          }
        }
      }
    },
    "/Center/{id}/units/{unitCode}/code": {
      "get": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_GetUnitByCode",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "unitCode",
            "in": "path",
            "required": true,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetUnitDto"
            }
          }
        }
      }
    },
    "/Center/{id}/units/leases/{leaseId}": {
      "get": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_GetLeaseById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "leaseId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetLeaseDto"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_DeleteLeaseById",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "leaseId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Lease"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_UpdateLeaseInCenter",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "leaseId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "leaseDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateLeaseDto"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Lease"
            }
          }
        }
      }
    },
    "/Center/{id}/units/{unitId}/leases/valid": {
      "get": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_GetValidLeaseByUnitId",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "unitId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetLeaseDto"
            }
          }
        }
      }
    },
    "/Center/{id}/units/leases": {
      "get": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_ListLeasesInCenter",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetLeaseDto"
              }
            }
          }
        }
      }
    },
...
    "/Center/{id}/units/{unitId}/leases": {
      "post": {
        "tags": [
          "Center"
        ],
        "operationId": "Center_CreateLeaseInCenter",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "unitId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "name": "leaseDto",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateLeaseDto"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Lease"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "GetCenterDto": {
      "type": "object",
      "required": [
        "id",
        "centerName",
        "owner",
        "parkingCapacity",
        "description"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "centerName": {
          "type": "string"
        },
        "owner": {
          "type": "string"
        },
        "parkingCapacity": {
          "type": "integer",
          "format": "int32"
        },
        "description": {
          "type": "string"
        }
      }
    },
    "Center": {
      "type": "object",
      "required": [
        "id",
        "centerName",
        "companyId",
        "owner",
        "floors",
        "parkingCapacity",
        "units",
        "description"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "centerName": {
          "type": "string"
        },
        "companyId": {
          "type": "integer",
          "format": "int32"
        },
        "owner": {
          "$ref": "#/definitions/Company"
        },
        "floors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Floor"
          }
        },
        "parkingCapacity": {
          "type": "integer",
          "format": "int32"
        },
        "units": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Unit"
          }
        },
        "description": {
          "type": "string"
        }
      }
    },
...
    "Unit": {
      "type": "object",
      "required": [
        "id",
        "unitCode",
        "area",
        "type",
        "centerId",
        "center",
        "floorId",
        "floor",
        "monthlyRentSqm",
        "monthlyMaintenanceCostSqm",
        "monthlyMarketingFeeSqm",
        "leases"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "unitCode": {
          "type": "string"
        },
        "area": {
          "type": "number",
          "format": "double"
        },
        "type": {
          "$ref": "#/definitions/UnitType"
        },
        "centerId": {
          "type": "integer",
          "format": "int32"
        },
        "center": {
          "$ref": "#/definitions/Center"
        },
        "floorId": {
          "type": "integer",
          "format": "int32"
        },
        "floor": {
          "$ref": "#/definitions/Floor"
        },
        "monthlyRentSqm": {
          "type": "number",
          "format": "double"
        },
        "monthlyMaintenanceCostSqm": {
          "type": "number",
          "format": "double"
        },
        "monthlyMarketingFeeSqm": {
          "type": "number",
          "format": "double"
        },
        "leases": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Lease"
          }
        }
      }
    },
    "UnitType": {
      "type": "integer",
      "description": "",
      "x-enumNames": [
        "Retail",
        "Storage",
        "Office"
      ],
      "enum": [
        0,
        1,
        2
      ]
    },
    "Floor": {
      "type": "object",
      "required": [
        "id",
        "floorName"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "floorName": {
          "type": "string"
        }
      }
    },
...
    "CreateCenterDto": {
      "type": "object",
      "required": [
        "centerName",
        "companyId",
        "parkingCapacity"
      ],
      "properties": {
        "centerName": {
          "type": "string"
        },
        "companyId": {
          "type": "integer",
          "format": "int32"
        },
        "parkingCapacity": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "UpdateCenterDto": {
      "type": "object",
      "required": [
        "centerName",
        "companyId",
        "parkingCapacity"
      ],
      "properties": {
        "centerName": {
          "type": "string"
        },
        "companyId": {
          "type": "integer",
          "format": "int32"
        },
        "parkingCapacity": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "GetUnitDto": {
      "type": "object",
      "required": [
        "id",
        "unitCode",
        "area",
        "type",
        "center",
        "floor",
        "monthlyRentSqm",
        "monthlyMaintenanceCostSqm",
        "monthlyMarketingFeeSqm"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "unitCode": {
          "type": "string"
        },
        "area": {
          "type": "number",
          "format": "double"
        },
        "type": {
          "type": "string"
        },
        "center": {
          "type": "string"
        },
        "floor": {
          "type": "string"
        },
        "monthlyRentSqm": {
          "type": "number",
          "format": "double"
        },
        "monthlyMaintenanceCostSqm": {
          "type": "number",
          "format": "double"
        },
        "monthlyMarketingFeeSqm": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "CreateUnitDto": {
      "type": "object",
      "required": [
        "unitCode",
        "area",
        "type",
        "floorId",
        "centerId",
        "monthlyRentSqm",
        "monthlyMaintenanceCostSqm",
        "monthlyMarketingFeeSqm"
      ],
      "properties": {
        "unitCode": {
          "type": "string"
        },
        "area": {
          "type": "number",
          "format": "double"
        },
        "type": {
          "type": "integer",
          "format": "int32"
        },
        "floorId": {
          "type": "integer",
          "format": "int32"
        },
        "centerId": {
          "type": "integer",
          "format": "int32"
        },
        "monthlyRentSqm": {
          "type": "number",
          "format": "double"
        },
        "monthlyMaintenanceCostSqm": {
          "type": "number",
          "format": "double"
        },
        "monthlyMarketingFeeSqm": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "UpdateUnitDto": {
      "type": "object",
      "required": [
        "area",
        "monthlyRentSqm",
        "monthlyMaintenanceCostSqm",
        "monthlyMarketingFeeSqm"
      ],
      "properties": {
        "area": {
          "type": "number",
          "format": "double"
        },
        "monthlyRentSqm": {
          "type": "number",
          "format": "double"
        },
        "monthlyMaintenanceCostSqm": {
          "type": "number",
          "format": "double"
        },
        "monthlyMarketingFeeSqm": {
          "type": "number",
          "format": "double"
        }
      }
    },
...
  }
}